pyproject.toml
[tool.poetry]
name = "simple-ddpm"
version = "0.1.0"
description = ""
authors = ["Cristian Garcia <cgarcia.e88@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8,<3.11"
jax = ">=0.3.17"
jaxlib = ">=0.3.15"
ml-collections = ">=0.1.1"
scikit-learn = ">=1.1.2"
matplotlib = ">=3.5.3"
flax = ">=0.6.0"
einop = ">=0.0.1"
pre-commit = ">=2.20.0"
jupytext = ">=1.14.1"
nbconvert = ">=6.2.0,<7.0.0"
jupyter-contrib-nbextensions = ">=0.5.1"
datasets = ">=2.4.0"
pkbar = ">=0.5"
tensorflow-cpu = ">=2.10.0"
wandb = ">=0.13.3"
gitignore-parser = ">=0.1.0"
tensorflow-datasets = ">=4.7.0"
moviepy = ">=1.0.3"
imageio = ">=2.22.2"
ciclo = {path = "../ciclo", develop = true}


[tool.poetry.dev-dependencies]
black = {version = "^22.8.0", allow-prereleases = true}
ipykernel = "^6.15.2"

[tool.poetry.group.dev.dependencies]
mkdocs-material = "^8.5.6"
mkdocs-jupyter = "^0.22.0"
black = {version = "^22.8.0", allow-prereleases = true}
debugpy = "^1.6.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
